requirements.txt
gradio>=4.0.0
transformers>=4.40.0
torch>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0